+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added `quantile.type` argument to `stats.default` to allow selecting which quantile type is computed.
 
+* Special handling for "0%" and "100%" in `stats.apply.rounding` (no decimals).
+
 # table1 1.0
 
 * This is the initial release of table1 on CRAN.